final commit

+++ web-services/db-access/src/main/java/io/github/tastypenguinbacon/nordic/db/resource/DataAccess.java
@@ -32,7 +32,7 @@ public class DataAccess {
     @Produces(MediaType.APPLICATION_JSON)
     @Path("{station}")
     public List<JsonNode> getData(@PathParam("station") @NotNull String station,
-                                @QueryParam("data-after") @DefaultValue("0") int dataAfter) {
+                                @QueryParam("data-after") @DefaultValue("0") long dataAfter) {
         return dataBase.getAfter(station, dataAfter).toJavaList();
     }
